+++ source/src/FlyCameraMissionView.cs
@@ -1,17 +1,13 @@
 ﻿using System;
 using System.Linq;
-using TaleWorlds.CampaignSystem.Conversation.Tags;
 using TaleWorlds.Core;
-using TaleWorlds.DotNet;
 using TaleWorlds.Engine;
 using TaleWorlds.Engine.Screens;
 using TaleWorlds.InputSystem;
 using TaleWorlds.Library;
 using TaleWorlds.MountAndBlade;
-using TaleWorlds.MountAndBlade.GauntletUI.Mission;
 using TaleWorlds.MountAndBlade.View;
 using TaleWorlds.MountAndBlade.View.Missions;
-using MathF = TaleWorlds.Library.MathF;
 
 namespace RTSCamera
 {
@@ -29,17 +25,17 @@ namespace RTSCamera
         private float _cameraHeightToAdd;
         private float _cameraHeightLimit;
         private bool _classicMode = true;
-        private bool _isOrderViewOpen = false;
+        private bool _isOrderViewOpen;
         private bool _resetDraggingMode;
         private bool _rightButtonDraggingMode;
         private Vec2 _clickedPositionPixel = Vec2.Zero;
-        private bool _setCombatActionOnNextTick = false;
-        private bool _levelToEdge = false;
-        private bool _lockToAgent = false;
+        private bool _setCombatActionOnNextTick;
+        private bool _levelToEdge;
+        private bool _lockToAgent;
 
-        private bool _forceMove = false;
+        private bool _forceMove;
         private Vec3 _forceMoveVec;
-        private float _forceMoveInvertedProgress = 0;
+        private float _forceMoveInvertedProgress;
 
         private float _cameraBearingDelta;
         private float _cameraElevationDelta;
@@ -50,6 +46,13 @@ namespace RTSCamera
         public float CameraElevation { get; private set; }
 
         public bool CameraRotateSmoothMode = false;
+
+        public bool ConstantSpeed = false;
+
+        public bool Outdoor = true;
+
+        public bool RestrictByBoundaries = true;
+
         public float CameraViewAngle
         {
             get => _cameraViewAngle;
@@ -112,9 +115,9 @@ namespace RTSCamera
         public float CameraSpeedFactor = 1;
 
         public float CameraVerticalSpeedFactor = 1;
-        private float _depthOfFieldDistance = 0;
-        private float _depthOfFieldStart = 0;
-        private float _depthOfFieldEnd = 0;
+        private float _depthOfFieldDistance;
+        private float _depthOfFieldStart;
+        private float _depthOfFieldEnd;
 
         //public float CameraBearing => MissionScreen.CameraBearing;
 
@@ -131,14 +134,17 @@ namespace RTSCamera
         public override void OnMissionScreenInitialize()
         {
             base.OnMissionScreenInitialize();
-            this._cameraSpeed = Vec3.Zero;
-            this._cameraSpeedMultiplier = 1f;
-            this._cameraHeightToAdd = 0.0f;
-            this._cameraHeightLimit = 0.0f;
-            this._cameraSmoothMode = true;
-            this.ViewOrderPriorty = 25;
+            _cameraSpeed = Vec3.Zero;
+            _cameraSpeedMultiplier = 1f;
+            _cameraHeightToAdd = 0.0f;
+            _cameraHeightLimit = 0.0f;
+            _cameraSmoothMode = true;
+            ViewOrderPriorty = 25;
 
             _config = RTSCameraConfig.Get();
+            ConstantSpeed = _config.ConstantSpeed;
+            Outdoor = _config.Outdoor;
+            RestrictByBoundaries = _config.RestrictByBoundaries;
             _freeCameraLogic = Mission.GetMissionBehaviour<SwitchFreeCameraLogic>();
             _missionMainAgentController = Mission.GetMissionBehaviour<MissionMainAgentController>();
             _orderUIHandler = Mission.GetMissionBehaviour<RTSCameraOrderUIHandler>();
@@ -183,8 +189,8 @@ namespace RTSCamera
         public SpectatorCameraTypes GetMissionCameraLockMode(bool lockedToMainPlayer)
         {
             ICameraModeLogic otherCameraModeLogic =
-                this.Mission.MissionBehaviours.FirstOrDefault<MissionBehaviour>(
-                        (Func<MissionBehaviour, bool>)(b => !(b is FlyCameraMissionView) && b is ICameraModeLogic)) as
+                Mission.MissionBehaviours.FirstOrDefault(
+                        b => !(b is FlyCameraMissionView) && b is ICameraModeLogic) as
                     ICameraModeLogic;
             return _lockToAgent && (_freeCameraLogic?.isSpectatorCamera ?? false)
                 ? SpectatorCameraTypes.LockToAnyAgent
@@ -222,35 +228,35 @@ namespace RTSCamera
                 _missionMainAgentController.IsDisabled = freeCamera;
             }
 
-            this._lockToAgent = false;
+            _lockToAgent = false;
             if (freeCamera)
             {
-                this.CameraBearing = MissionScreen.CameraBearing;
-                this.CameraElevation = MissionScreen.CameraElevation;
+                CameraBearing = MissionScreen.CameraBearing;
+                CameraElevation = MissionScreen.CameraElevation;
                 BeginForcedMove(new Vec3(0, 0, _config.RaisedHeight));
             }
         }
 
         private void UpdateDragData()
         {
-            if (this._resetDraggingMode)
+            if (_resetDraggingMode)
             {
-                this._rightButtonDraggingMode = false;
-                this._resetDraggingMode = false;
+                _rightButtonDraggingMode = false;
+                _resetDraggingMode = false;
                 EndDrag();
             }
             else if (_rightButtonDraggingMode && MissionScreen.SceneLayer.Input.IsKeyReleased(InputKey.RightMouseButton))
-                this._resetDraggingMode = true;
+                _resetDraggingMode = true;
             else if (MissionScreen.SceneLayer.Input.IsKeyPressed(InputKey.RightMouseButton))
             {
-                this._clickedPositionPixel = MissionScreen.SceneLayer.Input.GetMousePositionPixel();
+                _clickedPositionPixel = MissionScreen.SceneLayer.Input.GetMousePositionPixel();
             }
             else
             {
-                if (!_isOrderViewOpen || !MissionScreen.SceneLayer.Input.IsKeyDown(InputKey.RightMouseButton) || MissionScreen.SceneLayer.Input.IsKeyReleased(InputKey.RightMouseButton) || ((double)MissionScreen.SceneLayer.Input.GetMousePositionPixel().DistanceSquared(this._clickedPositionPixel) <= 10.0 || this._rightButtonDraggingMode))
+                if (!_isOrderViewOpen || !MissionScreen.SceneLayer.Input.IsKeyDown(InputKey.RightMouseButton) || MissionScreen.SceneLayer.Input.IsKeyReleased(InputKey.RightMouseButton) || (MissionScreen.SceneLayer.Input.GetMousePositionPixel().DistanceSquared(_clickedPositionPixel) <= 10.0 || _rightButtonDraggingMode))
                     return;
                 BeginDrag();
-                this._rightButtonDraggingMode = true;
+                _rightButtonDraggingMode = true;
             }
         }
 
@@ -259,39 +265,52 @@ namespace RTSCamera
             HandleRotateInput(dt);
             MatrixFrame cameraFrame = MatrixFrame.Identity;
             cameraFrame.rotation.RotateAboutSide(1.570796f);
-            cameraFrame.rotation.RotateAboutForward(this.CameraBearing);
-            cameraFrame.rotation.RotateAboutSide(this.CameraElevation);
+            cameraFrame.rotation.RotateAboutForward(CameraBearing);
+            cameraFrame.rotation.RotateAboutSide(CameraElevation);
             cameraFrame.origin = CameraPosition;
             if (_forceMove)
                 cameraFrame.origin += ForcedMoveTick(dt);
-            float heightAtPosition = this.Mission.Scene.GetGroundHeightAtPosition(cameraFrame.origin, BodyFlags.CommonCollisionExcludeFlags, true);
-            float heightFactorForHorizontalMove = MathF.Clamp((float)(1.0 + ((double)cameraFrame.origin.z - (double)heightAtPosition - 0.5) / 2), 1, 30);
-            float heightFactorForVerticalMove = MathF.Clamp((float)(1.0 + ((double)cameraFrame.origin.z - (double)heightAtPosition - 0.5) / 2), 1, 20);
-            if (this.DebugInput.IsHotKeyPressed("MissionScreenHotkeyIncreaseCameraSpeed"))
-                this._cameraSpeedMultiplier *= 1.5f;
-            if (this.DebugInput.IsHotKeyPressed("MissionScreenHotkeyDecreaseCameraSpeed"))
-                this._cameraSpeedMultiplier *= 0.6666667f;
-            if (this.DebugInput.IsHotKeyPressed("ResetCameraSpeed"))
-                this._cameraSpeedMultiplier = 1f;
-            this._cameraSpeed *= (float)(1.0 - 10 * (double)dt);
-            if (this.DebugInput.IsControlDown())
+            float heightFactorForHorizontalMove;
+            float heightFactorForVerticalMove;
+            if(!ConstantSpeed)
+            {
+                float heightAtPosition =
+                    Mission.Scene.GetGroundHeightAtPosition(cameraFrame.origin, BodyFlags.CommonCollisionExcludeFlags, false);
+                heightFactorForHorizontalMove = MathF.Clamp((float) (1.0 + (cameraFrame.origin.z - (double) heightAtPosition - 0.5) / 2),
+                    1, 30);
+                heightFactorForVerticalMove = MathF.Clamp((float) (1.0 + (cameraFrame.origin.z - (double) heightAtPosition - 0.5) / 2),
+                    1, 20);
+            }
+            else
+            {
+                heightFactorForHorizontalMove = 1;
+                heightFactorForVerticalMove = 1;
+            }
+            if (DebugInput.IsHotKeyPressed("MissionScreenHotkeyIncreaseCameraSpeed"))
+                _cameraSpeedMultiplier *= 1.5f;
+            if (DebugInput.IsHotKeyPressed("MissionScreenHotkeyDecreaseCameraSpeed"))
+                _cameraSpeedMultiplier *= 0.6666667f;
+            if (DebugInput.IsHotKeyPressed("ResetCameraSpeed"))
+                _cameraSpeedMultiplier = 1f;
+            _cameraSpeed *= (float)(1.0 - 10 * (double)dt);
+            if (DebugInput.IsControlDown())
             {
                 float num = MissionScreen.SceneLayer.Input.GetDeltaMouseScroll() * 0.008333334f;
-                if ((double)num > 0.00999999977648258)
-                    this._cameraSpeedMultiplier *= 1.25f;
-                else if ((double)num < -0.00999999977648258)
-                    this._cameraSpeedMultiplier *= 0.8f;
+                if (num > 0.00999999977648258)
+                    _cameraSpeedMultiplier *= 1.25f;
+                else if (num < -0.00999999977648258)
+                    _cameraSpeedMultiplier *= 0.8f;
             }
-            float num1 = 3f * this._cameraSpeedMultiplier * CameraSpeedFactor;
+            float num1 = 3f * _cameraSpeedMultiplier * CameraSpeedFactor;
             if (MissionScreen.SceneLayer.Input.IsGameKeyDown(23))
-                num1 *= (float)this._shiftSpeedMultiplier;
-            if (!this._cameraSmoothMode)
+                num1 *= _shiftSpeedMultiplier;
+            if (!_cameraSmoothMode)
             {
-                this._cameraSpeed.x = 0.0f;
-                this._cameraSpeed.y = 0.0f;
-                this._cameraSpeed.z = 0.0f;
+                _cameraSpeed.x = 0.0f;
+                _cameraSpeed.y = 0.0f;
+                _cameraSpeed.z = 0.0f;
             }
-            if (!this.DebugInput.IsControlDown() || !this.DebugInput.IsAltDown())
+            if (!DebugInput.IsControlDown() || !DebugInput.IsAltDown())
             {
                 Vec3 keyInput = Vec3.Zero;
                 Vec3 mouseInput = Vec3.Zero;
@@ -328,27 +347,27 @@ namespace RTSCamera
                     keyInput.Normalize();
                 if (mouseInput.LengthSquared > 0.0)
                     mouseInput.Normalize();
-                this._cameraSpeed += (keyInput + mouseInput) * num1 * heightFactorForHorizontalMove;
+                _cameraSpeed += (keyInput + mouseInput) * num1 * heightFactorForHorizontalMove;
             }
             float horizontalLimit = heightFactorForHorizontalMove * num1;
             float verticalLimit = heightFactorForVerticalMove * num1 * CameraVerticalSpeedFactor;
-            this._cameraSpeed.x = MBMath.ClampFloat(this._cameraSpeed.x, -horizontalLimit, horizontalLimit);
-            this._cameraSpeed.y = MBMath.ClampFloat(this._cameraSpeed.y, -horizontalLimit, horizontalLimit);
-            this._cameraSpeed.z = MBMath.ClampFloat(this._cameraSpeed.z, -verticalLimit, verticalLimit);
+            _cameraSpeed.x = MBMath.ClampFloat(_cameraSpeed.x, -horizontalLimit, horizontalLimit);
+            _cameraSpeed.y = MBMath.ClampFloat(_cameraSpeed.y, -horizontalLimit, horizontalLimit);
+            _cameraSpeed.z = MBMath.ClampFloat(_cameraSpeed.z, -verticalLimit, verticalLimit);
             if (_classicMode)
             {
                 Vec2 asVec2 = cameraFrame.origin.AsVec2;
-                cameraFrame.origin += this._cameraSpeed.x * new Vec3(cameraFrame.rotation.s.AsVec2, 0.0f, -1f).NormalizedCopy() * dt;
+                cameraFrame.origin += _cameraSpeed.x * cameraFrame.rotation.s.AsVec2.ToVec3().NormalizedCopy() * dt;
                 ref Vec3 local = ref cameraFrame.origin;
                 Vec3 vec3_2 = local;
-                double y = (double)this._cameraSpeed.y;
-                Vec3 vec3_1 = new Vec3(cameraFrame.rotation.u.AsVec2, 0.0f, -1f);
+                double y = _cameraSpeed.y;
+                Vec3 vec3_1 = cameraFrame.rotation.u.AsVec2.ToVec3();
                 Vec3 vec3_3 = vec3_1.NormalizedCopy();
                 Vec3 vec3_4 = (float)y * vec3_3 * dt;
                 local = vec3_2 - vec3_4;
-                cameraFrame.origin.z += this._cameraSpeed.z * dt;
+                cameraFrame.origin.z += _cameraSpeed.z * dt;
                 if (!MissionScreen.SceneLayer.Input.IsControlDown())
-                    this._cameraHeightToAdd -= (float)(TaleWorlds.InputSystem.Input.DeltaMouseScroll / 200.0) * verticalLimit;
+                    _cameraHeightToAdd -= (float)(TaleWorlds.InputSystem.Input.DeltaMouseScroll / 200.0) * verticalLimit;
                 // hold middle button and move mouse vertically to adjust height
                 if (MissionScreen.SceneLayer.Input.IsHotKeyDown("DeploymentCameraIsActive"))
                 {
@@ -357,43 +376,43 @@ namespace RTSCamera
                         _levelToEdge = true;
                         ScreenManager.FirstHitLayer.InputRestrictions.SetMouseVisibility(true);
                     }
-                    this._cameraHeightToAdd += 0.5f * TaleWorlds.InputSystem.Input.MouseMoveY;
+                    _cameraHeightToAdd += 0.5f * TaleWorlds.InputSystem.Input.MouseMoveY;
                 }
-                else if (_levelToEdge == true)
+                else if (_levelToEdge)
                 {
                     _levelToEdge = false;
                     ScreenManager.FirstHitLayer.InputRestrictions.SetMouseVisibility(false);
                 }
-                this._cameraHeightToAdd = MathF.Clamp(this._cameraHeightToAdd, -verticalLimit, verticalLimit);
-                if ((double)MathF.Abs(this._cameraHeightToAdd) > 1.0 / 1000.0)
+                _cameraHeightToAdd = MathF.Clamp(_cameraHeightToAdd, -verticalLimit, verticalLimit);
+                if (MathF.Abs(_cameraHeightToAdd) > 1.0 / 1000.0)
                 {
-                    cameraFrame.origin.z += (float)((double)this._cameraHeightToAdd * (double)dt);
-                    this._cameraHeightToAdd *= (float)(1.0 - dt * 5.0);
+                    cameraFrame.origin.z += (float)(_cameraHeightToAdd * (double)dt);
+                    _cameraHeightToAdd *= (float)(1.0 - dt * 5.0);
                 }
                 else
                 {
-                    cameraFrame.origin.z += this._cameraHeightToAdd * dt;
-                    this._cameraHeightToAdd = 0.0f;
+                    cameraFrame.origin.z += _cameraHeightToAdd * dt;
+                    _cameraHeightToAdd = 0.0f;
                 }
-                if ((double)this._cameraHeightLimit > 0.0 && (double)cameraFrame.origin.z > (double)this._cameraHeightLimit)
-                    cameraFrame.origin.z = this._cameraHeightLimit;
+                if (_cameraHeightLimit > 0.0 && cameraFrame.origin.z > (double)_cameraHeightLimit)
+                    cameraFrame.origin.z = _cameraHeightLimit;
             }
             else
             {
-                cameraFrame.origin += this._cameraSpeed.x * cameraFrame.rotation.s * dt;
-                cameraFrame.origin -= this._cameraSpeed.y * cameraFrame.rotation.u * dt;
-                cameraFrame.origin += this._cameraSpeed.z * cameraFrame.rotation.f * dt;
+                cameraFrame.origin += _cameraSpeed.x * cameraFrame.rotation.s * dt;
+                cameraFrame.origin -= _cameraSpeed.y * cameraFrame.rotation.u * dt;
+                cameraFrame.origin += _cameraSpeed.z * cameraFrame.rotation.f * dt;
             }
             if (!MBEditor.IsEditModeOn)
             {
-                if (!this.Mission.IsPositionInsideBoundaries(cameraFrame.origin.AsVec2))
-                    cameraFrame.origin.AsVec2 = this.Mission.GetClosestBoundaryPosition(cameraFrame.origin.AsVec2);
-                float heightAtPosition1 = this.Mission.Scene.GetGroundHeightAtPosition(cameraFrame.origin + new Vec3(0.0f, 0.0f, 100f, -1f), BodyFlags.CommonCollisionExcludeFlags, true);
-                if (!MissionScreen.IsCheatGhostMode && (double)heightAtPosition1 < 9999.0)
+                if (RestrictByBoundaries && !Mission.IsPositionInsideBoundaries(cameraFrame.origin.AsVec2))
+                    cameraFrame.origin.AsVec2 = Mission.GetClosestBoundaryPosition(cameraFrame.origin.AsVec2);
+                float heightAtPosition1 = Mission.Scene.GetGroundHeightAtPosition(cameraFrame.origin + new Vec3(0.0f, 0.0f, 100f));
+                if (!MissionScreen.IsCheatGhostMode && Outdoor && heightAtPosition1 < 9999.0)
                     cameraFrame.origin.z = Math.Max(cameraFrame.origin.z, heightAtPosition1 + 0.5f);
-                if ((double)cameraFrame.origin.z > (double)heightAtPosition1 + 80.0)
+                if (cameraFrame.origin.z > heightAtPosition1 + 80.0)
                     cameraFrame.origin.z = heightAtPosition1 + 80f;
-                if ((double)cameraFrame.origin.z < -100.0)
+                if (cameraFrame.origin.z < -100.0)
                     cameraFrame.origin.z = -100f;
             }
             UpdateCameraFrameAndDof(cameraFrame);
@@ -403,9 +422,9 @@ namespace RTSCamera
         {
             UpdateDof();
             UpdateViewAngle();
-            this.CombatCamera.Frame = matrixFrame;
-            MissionScreen.SceneView?.SetCamera(this.CombatCamera);
-            this.Mission.SetCameraFrame(matrixFrame, Zoom);
+            CombatCamera.Frame = matrixFrame;
+            MissionScreen.SceneView?.SetCamera(CombatCamera);
+            Mission.SetCameraFrame(matrixFrame, Zoom);
         }
 
         private void UpdateDof()
@@ -413,9 +432,9 @@ namespace RTSCamera
             Mission.Scene.SetDepthOfFieldParameters(DepthOfFieldStart, DepthOfFieldEnd, false);
             if (Math.Abs(DepthOfFieldDistance) < 0.0001f)
             {
-                this.Mission.Scene.RayCastForClosestEntityOrTerrain(this.CombatCamera.Position, this.CombatCamera.Position + this.CombatCamera.Direction * 3000f, out var terrainCollisionDistance, out GameEntity _, 0.5f, BodyFlags.CameraCollisionRayCastExludeFlags);
-                this.Mission.RayCastForClosestAgent(this.CameraPosition,
-                    this.CombatCamera.Position + this.CombatCamera.Direction * 3000f, out var agentCollisionDistance);
+                Mission.Scene.RayCastForClosestEntityOrTerrain(CombatCamera.Position, CombatCamera.Position + CombatCamera.Direction * 3000f, out var terrainCollisionDistance, out GameEntity _, 0.5f, BodyFlags.CameraCollisionRayCastExludeFlags);
+                Mission.RayCastForClosestAgent(CameraPosition,
+                    CombatCamera.Position + CombatCamera.Direction * 3000f, out var agentCollisionDistance);
                 if (float.IsNaN(terrainCollisionDistance))
                 {
                     terrainCollisionDistance = float.MaxValue;
@@ -425,7 +444,7 @@ namespace RTSCamera
                 {
                     agentCollisionDistance = float.MaxValue;
                 }
-                this.Mission.Scene.SetDepthOfFieldFocus(Math.Min(terrainCollisionDistance, agentCollisionDistance));
+                Mission.Scene.SetDepthOfFieldFocus(Math.Min(terrainCollisionDistance, agentCollisionDistance));
             }
             else
             {
@@ -435,8 +454,8 @@ namespace RTSCamera
 
         private void UpdateViewAngle()
         {
-            float newDNear = !this.Mission.CameraIsFirstPerson ? 0.1f : 0.065f;
-            this.CombatCamera.SetFovVertical((float)(this.CameraViewAngle * (Math.PI / 180.0)), TaleWorlds.Engine.Screen.AspectRatio, newDNear, 12500f);
+            float newDNear = !Mission.CameraIsFirstPerson ? 0.1f : 0.065f;
+            CombatCamera.SetFovVertical((float)(CameraViewAngle * (Math.PI / 180.0)), Screen.AspectRatio, newDNear, 12500f);
         }
 
         private void BeginForcedMove(Vec3 vec)
@@ -449,7 +468,7 @@ namespace RTSCamera
         private Vec3 ForcedMoveTick(float dt)
         {
             var previousProgress = _forceMoveInvertedProgress;
-            _forceMoveInvertedProgress *= (float)Math.Pow(0.00001, (double)dt);
+            _forceMoveInvertedProgress *= (float)Math.Pow(0.00001, dt);
             if (Math.Abs(_forceMoveInvertedProgress) < 0.0001f)
             {
                 _forceMove = false;
@@ -465,7 +484,7 @@ namespace RTSCamera
             float mouseSensitivity = MissionScreen.SceneLayer.Input.GetMouseSensivity();
             float inputXRaw = 0.0f;
             float inputYRaw = 0.0f;
-            if (!MBCommon.IsPaused && this.Mission.Mode != MissionMode.Barter)
+            if (!MBCommon.IsPaused && Mission.Mode != MissionMode.Barter)
             {
                 if (MissionScreen.MouseVisible && !MissionScreen.SceneLayer.Input.IsKeyDown(InputKey.RightMouseButton))
                 {
@@ -480,7 +499,7 @@ namespace RTSCamera
                 {
                     if (!MissionScreen.SceneLayer.Input.GetIsMouseActive())
                     {
-                        double num3 = (double)dt / 0.000600000028498471;
+                        double num3 = dt / 0.000600000028498471;
                         inputXRaw = (float)num3 * MissionScreen.SceneLayer.Input.GetGameKeyAxis("CameraAxisX") +
                                     MissionScreen.SceneLayer.Input.GetMouseMoveX();
                         inputYRaw = (float)-num3 * MissionScreen.SceneLayer.Input.GetGameKeyAxis("CameraAxisY") +
@@ -495,37 +514,37 @@ namespace RTSCamera
             }
             float num4 = 5.4E-05f;
             float smoothFading;
-            if (this.CameraRotateSmoothMode)
+            if (CameraRotateSmoothMode)
             {
                 num4 *= 0.10f;
                 smoothFading = (float)Math.Pow(0.000001, dt);
             }
             else
                 smoothFading = 0.0f;
-            this._cameraBearingDelta *= smoothFading;
-            this._cameraElevationDelta *= smoothFading;
+            _cameraBearingDelta *= smoothFading;
+            _cameraElevationDelta *= smoothFading;
             bool isSessionActive = GameNetwork.IsSessionActive;
             float inputScale = num4 * mouseSensitivity * MissionScreen.CameraViewAngle;
             float inputX = -inputXRaw * inputScale;
             float inputY = (NativeConfig.InvertMouse ? inputYRaw : -inputYRaw) * inputScale;
             if (isSessionActive)
             {
-                float maxValue = (float)(0.300000011920929 + 10.0 * (double)dt);
+                float maxValue = (float)(0.300000011920929 + 10.0 * dt);
                 inputX = MBMath.ClampFloat(inputX, -maxValue, maxValue);
                 inputY = MBMath.ClampFloat(inputY, -maxValue, maxValue);
             }
-            this._cameraBearingDelta += inputX;
-            this._cameraElevationDelta += inputY;
+            _cameraBearingDelta += inputX;
+            _cameraElevationDelta += inputY;
             if (isSessionActive)
             {
-                float maxValue = (float)(0.300000011920929 + 10.0 * (double)dt);
-                this._cameraBearingDelta = MBMath.ClampFloat(this._cameraBearingDelta, -maxValue, maxValue);
-                this._cameraElevationDelta = MBMath.ClampFloat(this._cameraElevationDelta, -maxValue, maxValue);
+                float maxValue = (float)(0.300000011920929 + 10.0 * dt);
+                _cameraBearingDelta = MBMath.ClampFloat(_cameraBearingDelta, -maxValue, maxValue);
+                _cameraElevationDelta = MBMath.ClampFloat(_cameraElevationDelta, -maxValue, maxValue);
             }
 
-            this.CameraBearing += this._cameraBearingDelta;
-            this.CameraElevation += this._cameraElevationDelta;
-            this.CameraElevation = MBMath.ClampFloat(this.CameraElevation, -1.36591f, 1.121997f);
+            CameraBearing += _cameraBearingDelta;
+            CameraElevation += _cameraElevationDelta;
+            CameraElevation = MBMath.ClampFloat(CameraElevation, -1.36591f, 1.121997f);
         }
 
         public override void OnPreDisplayMissionTick(float dt)
@@ -536,7 +555,7 @@ namespace RTSCamera
                 UpdateDragData();
             if (!_isOrderViewOpen &&
                 (Input.IsGameKeyReleased(8) ||
-                 (Input.IsGameKeyReleased(9) && !this._rightButtonDraggingMode)))
+                 (Input.IsGameKeyReleased(9) && !_rightButtonDraggingMode)))
             {
                 _lockToAgent = true;
             }
@@ -548,8 +567,8 @@ namespace RTSCamera
                                   Input.GetKeyState(InputKey.ControllerLStick).x != 0.0)))
             {
                 _lockToAgent = false;
-                this.CameraBearing = MissionScreen.CameraBearing;
-                this.CameraElevation = MissionScreen.CameraElevation;
+                CameraBearing = MissionScreen.CameraBearing;
+                CameraElevation = MissionScreen.CameraElevation;
             }
         }
     }