+++ source/src/RTSCameraConfig.cs
@@ -6,11 +6,11 @@ namespace RTSCamera
 {
     public class RTSCameraConfig : RTSCameraConfigBase<RTSCameraConfig>
     {
-        protected static Version BinaryVersion => new Version(1, 0);
+        protected static Version BinaryVersion => new Version(1, 1);
 
         protected override void UpgradeToCurrentVersion()
         {
-            switch (ConfigVersion?.ToString())
+            switch (ConfigVersion)
             {
                 default:
                     Utility.DisplayLocalizedText("str_em_config_incompatible");
@@ -18,6 +18,11 @@ namespace RTSCamera
                     Serialize();
                     break;
                 case "1.0":
+                    ConstantSpeed = false;
+                    Outdoor = true;
+                    RestrictByBoundaries = true;
+                    break;
+                case "1.1":
                     break;
             }
         }
@@ -26,23 +31,29 @@ namespace RTSCamera
 
         public string ConfigVersion { get; set; } = BinaryVersion.ToString();
 
-        public bool UseFreeCameraByDefault = false;
+        public bool UseFreeCameraByDefault;
 
         public float RaisedHeight = 10;
 
         public int PlayerFormation = 4;
 
-        public bool DisableDeath = false;
+        public bool DisableDeath;
 
-        public bool SlowMotionMode = false;
+        public bool ConstantSpeed;
+
+        public bool Outdoor = true;
+
+        public bool RestrictByBoundaries = true;
+
+        public bool SlowMotionMode;
 
         public float SlowMotionFactor = 0.2f;
 
         public bool DisplayMessage = true;
 
-        public bool ControlAlliesAfterDeath = false;
+        public bool ControlAlliesAfterDeath;
 
-        public bool PreferToControlCompanions = false;
+        public bool PreferToControlCompanions;
 
         public bool ControlTroopsInPlayerPartyOnly = true;
 
@@ -65,17 +76,20 @@ namespace RTSCamera
 
         protected override void CopyFrom(RTSCameraConfig other)
         {
-            this.ConfigVersion = other.ConfigVersion;
-            this.UseFreeCameraByDefault = other.UseFreeCameraByDefault;
-            this.RaisedHeight = other.RaisedHeight;
-            this.PlayerFormation = other.PlayerFormation;
-            this.DisableDeath = other.DisableDeath;
-            this.SlowMotionMode = other.SlowMotionMode;
-            this.SlowMotionFactor = other.SlowMotionFactor;
-            this.DisplayMessage = other.DisplayMessage;
-            this.ControlAlliesAfterDeath = other.ControlAlliesAfterDeath;
-            this.PreferToControlCompanions = other.PreferToControlCompanions;
-            this.ControlTroopsInPlayerPartyOnly = other.ControlTroopsInPlayerPartyOnly;
+            ConfigVersion = other.ConfigVersion;
+            UseFreeCameraByDefault = other.UseFreeCameraByDefault;
+            RaisedHeight = other.RaisedHeight;
+            PlayerFormation = other.PlayerFormation;
+            DisableDeath = other.DisableDeath;
+            ConstantSpeed = other.ConstantSpeed;
+            Outdoor = other.Outdoor;
+            RestrictByBoundaries = other.RestrictByBoundaries;
+            SlowMotionMode = other.SlowMotionMode;
+            SlowMotionFactor = other.SlowMotionFactor;
+            DisplayMessage = other.DisplayMessage;
+            ControlAlliesAfterDeath = other.ControlAlliesAfterDeath;
+            PreferToControlCompanions = other.PreferToControlCompanions;
+            ControlTroopsInPlayerPartyOnly = other.ControlTroopsInPlayerPartyOnly;
         }
 
         public override void ResetToDefault()