+++ source/src/RTSCameraSubModule.cs
@@ -1,19 +1,28 @@
-﻿using HarmonyLib;
+﻿using System.Reflection;
+using HarmonyLib;
 using TaleWorlds.Core;
 using TaleWorlds.Library;
 using TaleWorlds.MountAndBlade;
+using Module = TaleWorlds.MountAndBlade.Module;
 
 namespace RTSCamera
 {
     public class RTSCameraSubModule : MBSubModuleBase
     {
+        private readonly Harmony _harmony = new Harmony("LeaveDetachmentPatch");
+
+        private readonly MethodInfo _original =
+            typeof(Formation).GetMethod("LeaveDetachment", BindingFlags.Instance | BindingFlags.NonPublic);
+
+        private readonly MethodInfo _prefix =
+            typeof(LeaveDetachmentPatch).GetMethod("Prefix", BindingFlags.Static | BindingFlags.Public);
         protected override void OnSubModuleLoad()
         {
             base.OnSubModuleLoad();
             RTSCameraExtension.Clear();
             Module.CurrentModule.GlobalTextManager.LoadGameTexts(BasePath.Name + "Modules/RTSCamera/ModuleData/module_strings.xml");
 
-            new Harmony("LeaveDetachmentPatch").PatchAll();
+            _harmony.Patch(_original, new HarmonyMethod(_prefix));
         }
 
         protected override void OnGameStart(Game game, IGameStarter gameStarterObject)
@@ -27,6 +36,7 @@ namespace RTSCamera
         {
             base.OnSubModuleUnloaded();
             RTSCameraExtension.Clear();
+            _harmony.UnpatchAll(_harmony.Id);
         }
     }
 }